+++ src/App.jsx
@@ -1,9 +1,8 @@
 import React, { useState } from 'react'
 import { createRoot } from 'react-dom/client'
 import { BrowserRouter, Routes, Route, Link } from 'react-router-dom'
-import { RobotsProvider } from './RobotsContext'
-import { Details } from './Details'
-import { Robots } from './Robots'
+import { RobotsProvider } from './context/index'
+import { Robots, Details } from './pages/index'
 import 'tachyons'
 import './index.css'
