+++ src/components/index.js
@@ -1,0 +1,5 @@
+export * from './Card'
+export * from './CardList'
+export * from './Modal'
+export * from './Pagination'
+export * from './SearchBox'
