+++ src/data/index.js
@@ -1,3 +1,2 @@
 export * from './jobList'
 export * from './seniorityList'
-export * from './functions'
