+++ src/data/seniorityList.js
@@ -1,5 +1,4 @@
 export const seniorityList = [
-  '',
   'Trainee',
   'Junior',
   'Mid-level',
