+++ src/pages/index.js
@@ -1,0 +1,2 @@
+export * from './Robots'
+export * from './Details'
