+++ src/utils/index.js
@@ -1,0 +1,2 @@
+export { fetchRobots } from './fetchRobots'
+export * from './functions'
